refact: rename platform.theme.ui package to 'kit'
