+++ platform/theme/src/main/java/kit/action/setting/group/SettingGroupScope.kt
@@ -1,11 +1,11 @@
-package com.jeanbarrossilva.orca.platform.theme.ui.action.setting.group
+package com.jeanbarrossilva.orca.platform.theme.kit.action.setting.group
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Modifier
-import com.jeanbarrossilva.orca.platform.theme.ui.action.setting.Setting
+import com.jeanbarrossilva.orca.platform.theme.kit.action.setting.Setting
 
 /** Scope of a [SettingGroup], through which [setting]s can be added. **/
 class SettingGroupScope internal constructor() {
