+++ platform/theme/src/main/java/kit/action/setting/group/SettingMetadata.kt
@@ -1,9 +1,9 @@
-package com.jeanbarrossilva.orca.platform.theme.ui.action.setting.group
+package com.jeanbarrossilva.orca.platform.theme.kit.action.setting.group
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.jeanbarrossilva.orca.platform.theme.ui.action.setting.Setting
+import com.jeanbarrossilva.orca.platform.theme.kit.action.setting.Setting
 import java.util.Objects
 
 /**
