+++ platform/theme/src/main/java/kit/menu/DropdownMenu.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.orca.platform.theme.ui.menu
+package com.jeanbarrossilva.orca.platform.theme.kit.menu
 
 import android.content.res.Configuration
 import androidx.compose.foundation.background
